Removing from list if pre-requisites are missing

+++ syllabus-recommendations/src/main/java/com/syllabus/data/model/StudentDataModel.java
@@ -27,7 +27,7 @@ public class StudentDataModel {
     private String userId;
 
     @Field(name = "courses_taken")
-    private Long coursesTaken;
+    private Integer coursesTaken;
 
     @Field(name = "term")
     private Short studentTerm;
